+++ .eslintignore
@@ -1,0 +1,3 @@
+dist/
+lib/
+node_modules/
