+++ action.yaml
@@ -10,9 +10,6 @@ inputs:
   appspec-file:
     description: 'Appspec file'
     required: true
-#outputs:
-#  time: # id of output
-#    description: 'The time we greeted you'
 runs:
   using: 'node12'
   main: 'dist/index.js'
