+++ package.json
@@ -2,10 +2,15 @@
   "name": "github-action-codedeploy",
   "version": "1.0.0",
   "description": "codedeploy github action",
-  "main": "index.js",
+  "main": "lib/main.js",
   "scripts": {
-    "package": "ncc build index.js -m -o dist",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc",
+    "format": "prettier --write **/*.ts",
+    "format-check": "prettier --check **/*.ts",
+    "lint": "eslint src/**/*.ts",
+    "pack": "ncc build",
+    "test": "jest",
+    "all": "yarn build && yarn format && yarn lint && yarn run pack && yarn test"
   },
   "author": "Scott Hutchings",
   "license": "Apache-2.0",
@@ -14,6 +19,19 @@
     "aws-sdk": "^2.677.0"
   },
   "devDependencies": {
-    "@zeit/ncc": "^0.22.3"
+    "@types/jest": "^24.0.23",
+    "@types/node": "^12.7.12",
+    "@typescript-eslint/parser": "^2.8.0",
+    "@zeit/ncc": "^0.20.5",
+    "eslint": "^5.16.0",
+    "eslint-plugin-github": "^2.0.0",
+    "eslint-plugin-jest": "^22.21.0",
+    "eslint-plugin-prettier": "^3.1.3",
+    "jest": "^24.9.0",
+    "jest-circus": "^24.9.0",
+    "js-yaml": "^3.13.1",
+    "prettier": "^1.19.1",
+    "ts-jest": "^24.2.0",
+    "typescript": "^3.6.4"
   }
 }
